extract all the node start procedure into init funtion
