+++ src/consensus_engine.rs
@@ -1,8 +1,18 @@
+use crate::bootstrap::{
+    init_mempool, init_network_interfaces, init_peers_and_metadata, start_consensus,
+};
+use crate::network::{create_network_runtime, extract_network_configs};
+use crate::storage::db::GravityDB;
 use crate::{GCEIError, GTxn, GravityConsensusEngineInterface};
+use aptos_config::config::NodeConfig;
+use aptos_config::network_id::NetworkId;
 use aptos_consensus::gravity_state_computer::ConsensusAdapterArgs;
 use aptos_crypto::hash::HashValue;
 use aptos_crypto::{PrivateKey, Uniform};
+use aptos_event_notifications::EventNotificationSender;
 use aptos_mempool::MempoolClientRequest;
+use aptos_network_builder::builder::NetworkBuilder;
+use aptos_storage_interface::DbReaderWriter;
 use aptos_types::account_address::AccountAddress;
 use aptos_types::chain_id::ChainId;
 use aptos_types::mempool_status::{MempoolStatus, MempoolStatusCode};
@@ -17,6 +27,7 @@ use futures::{
     SinkExt,
 };
 use std::collections::HashMap;
+use std::sync::Arc;
 use tokio::sync::{Mutex, RwLock};
 
 pub struct GravityConsensusEngine {
@@ -63,7 +74,87 @@ impl GravityConsensusEngine {
 
 #[async_trait::async_trait]
 impl GravityConsensusEngineInterface for GravityConsensusEngine {
-    fn init(args: &mut ConsensusAdapterArgs) -> Self {
+    fn init(node_config: NodeConfig, gravity_db: GravityDB) -> Self {
+        let peers_and_metadata = init_peers_and_metadata(&node_config, &gravity_db);
+        let db: DbReaderWriter = DbReaderWriter::new(gravity_db);
+        let mut event_subscription_service =
+            aptos_event_notifications::EventSubscriptionService::new(Arc::new(
+                aptos_infallible::RwLock::new(db.clone()),
+            ));
+        let network_configs = extract_network_configs(&node_config);
+
+        let network_config = network_configs.get(0).unwrap();
+        let chain_id = ChainId::test();
+        let mut network_builder = NetworkBuilder::create(
+            chain_id,
+            node_config.base.role,
+            &network_config,
+            aptos_time_service::TimeService::real(),
+            Some(&mut event_subscription_service),
+            peers_and_metadata.clone(),
+        );
+        let network_id: NetworkId = network_config.network_id;
+        let (consensus_network_interfaces, mempool_interfaces) = init_network_interfaces(
+            &mut network_builder,
+            network_id,
+            &network_config,
+            &node_config,
+            peers_and_metadata.clone(),
+        );
+        let state_sync_config = node_config.state_sync;
+        let (consensus_notifier, consensus_listener) =
+            aptos_consensus_notifications::new_consensus_notifier_listener_pair(
+                state_sync_config
+                    .state_sync_driver
+                    .commit_notification_timeout_ms,
+            );
+        let mut network_runtimes = vec![];
+        // Create a network runtime for the config
+        let runtime = create_network_runtime(&network_config);
+        // Build and start the network on the runtime
+        network_builder.build(runtime.handle().clone());
+        network_builder.start();
+        network_runtimes.push(runtime);
+
+        // TODO(Gravity_byteyue): delete the following comment
+        // 这里要看aptos代码的setup_environment_and_start_node函数下的start_mempool_runtime_and_get_consensus_sender的逻辑，不然这里channel好像对不上都
+        // start consensus确实是用consensus_to_mempool_receiver，但是在setup_environment_and_start_node才有Receiver<MempoolClientRequest>
+        // setup_environment_and_start_node 调用了 bootstrap_api_and_indexer ，在其中构造了 mempool_client_sender 和 mempool_client_receiver, 然后 bootstrap_api_and_indexer
+        // 返回了 receiver, 接下来 setup_environment_and_start_node 把 receiver 传递给 start_mempool_runtime_and_get_consensus_sender ,
+        // 在其中构造了 consensus_to_mempool_sender 和 consensus_to_mempool_receiver
+        // 并返回了sender
+        let (mempool_client_sender, mempool_client_receiver) = mpsc::channel(1);
+
+        let (consensus_to_mempool_sender, consensus_to_mempool_receiver) = mpsc::channel(1);
+        let (notification_sender, notification_receiver) = mpsc::channel(1);
+
+        let _mempool_notifier =
+            aptos_mempool_notifications::MempoolNotifier::new(notification_sender);
+        let mempool_listener =
+            aptos_mempool_notifications::MempoolNotificationListener::new(notification_receiver);
+
+        init_mempool(
+            &node_config,
+            &db,
+            &mut event_subscription_service,
+            mempool_client_sender.clone(),
+            mempool_interfaces,
+            mempool_client_receiver,
+            consensus_to_mempool_receiver,
+            mempool_listener,
+            peers_and_metadata,
+        );
+        let mut args = ConsensusAdapterArgs::new(mempool_client_sender);
+        start_consensus(
+            &node_config,
+            &mut event_subscription_service,
+            consensus_network_interfaces,
+            consensus_notifier,
+            consensus_to_mempool_sender,
+            db,
+            &args,
+        );
+        let _ = event_subscription_service.notify_initial_configs(1_u64);
         Self {
             mempool_sender: args.mempool_sender.clone(),
             pipeline_block_receiver: args.pipeline_block_receiver.take(),
@@ -138,7 +229,7 @@ impl GravityConsensusEngineInterface for GravityConsensusEngine {
                     gas_unit_price: txn.gas_unit_price(),
                     expiration_timestamp_secs: txn.expiration_timestamp_secs(),
                     chain_id: txn.chain_id().to_u8() as u64,
-                    txn_bytes: (*txn_bytes.clone()).to_owned()
+                    txn_bytes: (*txn_bytes.clone()).to_owned(),
                 }
             })
             .collect();
