+++ src/lib.rs
@@ -8,8 +8,9 @@ use aptos_config::config::NodeConfig;
 use aptos_consensus::gravity_state_computer::ConsensusAdapterArgs;
 use bootstrap::{check_bootstrap_config, start};
 use clap::Parser;
-use std::path::PathBuf;
+use storage::db::GravityDB;
 use std::fmt::Display;
+use std::path::PathBuf;
 
 pub struct GTxn {
     sequence_number: u64,
@@ -78,7 +79,7 @@ pub trait GravityConsensusEngineInterface: Send + Sync {
     /// - Setting up initial state
     /// - Connecting to the network
     /// - Loading configuration
-    fn init(args: &mut ConsensusAdapterArgs) -> Self;
+    fn init(node_config: NodeConfig, gravity_db: GravityDB) -> Self;
 
     /// Receive and process valid transactions.
     ///
